Refactor: constants 파일들 관리 개선
