+++ src/hooks/queries/comment/useCreateCommentMutation.ts
@@ -2,9 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { postComment, PostCommentParams } from '@api/post/comment/postComment';
 
-import { QUERY_KEYS } from '@constants/api';
-import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@constants/messages';
-import { STATUS_CODES } from '@constants/statusCodes';
+import { ERROR_MESSAGES, QUERY_KEYS, STATUS_CODES, SUCCESS_MESSAGES } from '@constants/index';
 
 // 댓글 생성
 export const useCreateCommentMutation = ({ postId }: { postId: string }) => {
