+++ src/hooks/queries/comment/useDeleteCommentMutation.ts
@@ -2,9 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { DeleteCommentParams, deleteComment } from '@api/post/comment/deleteComment';
 
-import { QUERY_KEYS } from '@constants/api';
-import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@constants/messages';
-import { STATUS_CODES } from '@constants/statusCodes';
+import { ERROR_MESSAGES, QUERY_KEYS, STATUS_CODES, SUCCESS_MESSAGES } from '@constants/index';
 
 // 댓글 삭제
 export const useDeleteCommentMutation = ({ postId }: { postId: string }) => {
