+++ src/hooks/queries/post/usePostsQuery.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query';
 
 import { getPosts, GetPostsParams } from '@api/post/getPosts';
 
-import { QUERY_KEYS } from '@constants/api';
+import { QUERY_KEYS } from '@constants/index';
 
 export const usePostsQuery = ({ page, limit }: GetPostsParams) => {
   return useQuery({
