+++ src/hooks/queries/user/useLogoutMutation.ts
@@ -2,8 +2,7 @@ import { useMutation } from '@tanstack/react-query';
 
 import { postLogout } from '@api/user/postLogout';
 
-import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@constants/messages';
-import { STATUS_CODES } from '@constants/statusCodes';
+import { ERROR_MESSAGES, STATUS_CODES, SUCCESS_MESSAGES } from '@constants/index';
 
 import { useAuthStore } from '@stores/useAuthStore';
